backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.3,<3.0
python-dotenv==1.0.0
google-generativeai==0.3.0
requests==2.31.0
httpx==0.27.0
apscheduler==3.10.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.109.0
email-validator==2.0.0
pandas==2.1.3
python-multipart==0.0.6
streamlit
pillow
huggingface_hub
tweepy
playwright
instagrapi
gradio_client
beautifulsoup4